Actualizacion del tema 3

+++ Primer trimestre/Tema 3/Ejercicio10.java
@@ -3,8 +3,8 @@ public class Ejercicio10 {
     String linea;
     System.out.print("Introduce la candidad de Mb que quiere convertir a Kb: ");
     linea = System.console().readLine();
-    int mb = Integer.parseInt( linea );
-    int kb = (mb * 8000);
+    double mb = Double.parseDouble(linea);
+    double kb = (mb * 8000);
     System.out.println("La cantidad añadida a Kb es: " + kb + "Kb");
   }
 }
