+++ Primer trimestre/Tema 3/Ejercicio11.java
@@ -3,8 +3,8 @@ public class Ejercicio11 {
     String linea;
     System.out.print("Introduce la candidad de Kb que quiere convertir a Mb: ");
     linea = System.console().readLine();
-    int kb = Integer.parseInt( linea );
-    int mb = (kb / 8000);
+    double kb = Double.parseDouble(linea);
+    double mb = (kb / 8000);
     System.out.println("La cantidad añadida convertida a Mb es: " + mb + "Mb");
   }
 }
