+++ Primer trimestre/Tema 3/Ejercicio8.java
@@ -3,8 +3,8 @@ public class Ejercicio8 {
     String linea;
     System.out.print("Introduce el número de horas que trabaja durante la semana: ");
     linea = System.console().readLine();
-    int horas = Integer.parseInt( linea );
-    int dinerosemanal;
+    double horas = Double.parseDouble(linea);
+    double dinerosemanal;
     dinerosemanal = (horas * 12);
     System.out.println("El sueldo corespondiente segun tus hoaras trabajadas a la semana es de: " + dinerosemanal + " euros");
   }
